\name{fit_glmm}
\alias{fit_glmm}
\title{Estimate conditional effects with generalized linear mixed-effects models}
\usage{
fit_glmm(data, formula, random_effects, family)
}
\description{
Estimate conditional effects with generalized linear mixed-effects models. Data must be in long format.
}
\examples{

}
